requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-cors-headers
mssql-django
python-dotenv
gunicorn
requests